删除反射验证值是否存在
